Moving recently created use cases inside v2 package
